+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/TigerConfigurationKeyString.java
@@ -35,4 +35,8 @@ public class TigerConfigurationKeyString {
     public String asLowerCase() {
         return value.toLowerCase();
     }
+
+    public String asString() {
+        return value;
+    }
 }