+++ ReleaseNotes.md
@@ -6,7 +6,7 @@
 * RestAssured 5.2.0
 * Selenium 4.8.0
 * Appium 8.3.0
-* **Spring Boot 3.0.1**
+* **Spring Boot 3.0.6**
 * Logback 1.2.11
 
 # Release 2.0.0
@@ -28,6 +28,18 @@
 * TGR-888: ${hostname} added as a configuration property
 * TGR-890: Workflow Ui now has a quit button in the left sidebar to abort the currently active test execution
 * TGR-890: Workflow Ui now has a pause button in the left sidebar to pause test execution. Please be aware that timeouts (e.g. while waiting for a rbel message to appear are not affected by the pause mode, thus using pause on these steps will cause a timeout and potentially subsequent test failure).
+* TGR-730: support changing log level similar to spring boot by specifying package or class name and level in tiger.yaml under logging.level
+```yaml
+servers:
+  server1:
+    ......
+logging:
+  level:
+    de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr: TRACE
+    de.gematik.test.tiger.lib.TigerDirector: TRACE
+    de.gematik.test.tiger.proxy: TRACE
+    localTigerProxy: TRACE
+```
 
 ## Breaking changes